rename plugin.h
