+++ source/core/plugin/Plugin.h
